requirements.txt
crewai>=0.28.0
streamlit>=1.29.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0